+++ beetsplug/lyrics.py
@@ -23,6 +23,7 @@ import unicodedata
 import urllib
 import difflib
 import itertools
+import warnings
 from HTMLParser import HTMLParseError
 
 from beets import logging
@@ -61,7 +62,13 @@ def fetch_url(url):
     is unreachable.
     """
     try:
-        r = requests.get(url, verify=False)
+        # Disable the InsecureRequestWarning that comes from using
+        # `verify=false`.
+        # https://github.com/kennethreitz/requests/issues/2214
+        # We're not overly worried about the NSA MITMing our lyrics scraper.
+        with warnings.catch_warnings():
+            warnings.simplefilter('ignore')
+            r = requests.get(url, verify=False)
     except requests.RequestException as exc:
         log.debug(u'lyrics request failed: {0}', exc)
         return